And finally (I think?), mic profile.

+++ ipc/src/lib.rs
@@ -27,7 +27,7 @@ pub enum GoXLRCommand {
     AssignFader(FaderName, ChannelName),
     SetVolume(ChannelName, u8),
     SetChannelMuted(ChannelName, bool, bool),
-    SetMicrophoneGain(MicrophoneType, u16),
+    SetMicrophoneGain(MicrophoneType, u8),
     LoadProfile(String),
     LoadMicProfile(String),
 }
